Ported to Racket.
class.scm has been replaced by a thin wrapper over Swindle. It's
hackish, and clunky in places, but it works.

The debugging console is known not to work, and it's not obvious how
to make it work properly with namespaces and all. I'm sure it can be
made to work, but it may take some thought.

A couple miscellaneous issues were fixed along the way. Copious
amounts of bugs have probably been introduced to compensate.